+++ internal/provider/services.go
@@ -7,6 +7,7 @@ import (
 
 	directory "google.golang.org/api/admin/directory/v1"
 	"google.golang.org/api/chromepolicy/v1"
+	"google.golang.org/api/cloudidentity/v1"
 	"google.golang.org/api/gmail/v1"
 	"google.golang.org/api/groupssettings/v1"
 )
@@ -43,6 +44,23 @@ func GetChromePolicySchemasService(chromePolicyService *chromepolicy.Service) (*
 	}
 
 	return customersService.PolicySchemas, diags
+}
+
+func GetCloudIdentityGroupsService(cloudIdentityService *cloudidentity.Service) (*cloudidentity.GroupsService, diag.Diagnostics) {
+	var diags diag.Diagnostics
+
+	log.Printf("[INFO] Instantiating Google Cloud Identity Groups service")
+	groupsService := cloudIdentityService.Groups
+	if groupsService == nil {
+		diags = append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  "Cloud Identity Groups Service could not be created.",
+		})
+
+		return nil, diags
+	}
+
+	return groupsService, diags
 }
 
 func GetDomainAliasesService(directoryService *directory.Service) (*directory.DomainAliasesService, diag.Diagnostics) {